pyproject.toml
[tool.poetry]
name = "GAAPF"
version = "0.1.0"
description = "GAAPF - Guidance AI Agent for Python Framework Learning"
authors = ["phamdinhkhanh <phamdinhkhanh.tkt53.neu@gmail.com>"]
readme = "README.md"
packages = [{include = "GAAPF", from = "src"}]

[tool.poetry.scripts]
GAAPF = "GAAPF.interfaces.cli.cli:main"

[tool.poetry.dependencies]
python = ">=3.10,<3.14"
langchain-core = "0.3.41"
langchain-openai = "0.3.7"
langchain-together = "0.3.0"
langgraph = "0.4.7"
python-dotenv = "1.0.0"
pandas = "2.2.3"
numpy = { version = "^1.26.0", optional = true }
matplotlib = "3.7.1"
plotly = "5.22.0"
yfinance = "0.2.54"
tavily-python = { version = "^0.3.1", optional = true }
googlenewsdecoder = "0.1.7"
aucodb = {extras = ["graph"], version = "0.1.7" }
mcp = "1.9.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.1"
black = "^24.4.2"
python-dotenv = "^1.0.0"
ipykernel = "^6.29.5"

[tool.poetry.group.test.dependencies]
pytest = "^8.2.2"
pytest-cov = "^6.0.0"
pandas = "2.2.3"
numpy = { version = "^1.26.0", optional = true }
matplotlib = "3.7.1"
plotly = "5.22.0"
yfinance = "0.2.54"
tavily-python = { version = "^0.3.1", optional = true }
googlenewsdecoder = "0.1.7"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.black]
line-length = 88
target-version = ["py310"]
include = '\.pyi?$'
extend-exclude = '''
/(
    \.git
    | \.venv
    | build
    | dist
    | data
    | \.pytest_cache
)/
'''

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
